Fix geo_id type in read_geos...() function family
Problem:
Type of 'geo_id' variable in geos datasets is not set;
it is sometimes of type integer and other times of type character.
In non-geos datasets (demographics, land area, etc.) 'geo_id' is
always of type character.

Fix:
Update all functions to coerce 'geo_id' to character when read.